+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## [unreleased]
+
+- Add label accessors (`----:com.apple.iTunes:LABEL`)
+
 ## mp4ameta v0.12.1
 
 - Add ReadConfig::NONE and WriteConfig::NONE